+++ packages/interfaces/CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Renamed `setup` to `start` and changed its return type.
+- Renamed `teardown` to `stop` and changed its return type.
 
 ## [4.0.0-alpha.1]
 ### Changed